moved timemanager into feature package
